refactor(audit): share median and clampToWindow across layers
Three median implementations (collectors/incidents.ts, metrics/_score.ts,
and a third in met-mttr-incidents.test.ts) and two window clamps —
clampIncidentsToWindow was a near-line-for-line twin of
metrics/_base.ts's generic clampToWindow, as its own doc comment said,
and the two had already drifted by the no-window short-circuit.

The cause was structural: collectors cannot import from metrics/ without
inverting the layering, since metrics/mttr.ts imports from
collectors/incidents.ts. So the shared helpers move to
collectors/_base.ts, which both layers can reach, and metrics/_base.ts
and metrics/_score.ts re-export them so no call site changes.

clampIncidentsToWindow becomes a thin wrapper and inherits the generic
version's tests, including the keep-records-with-no-parseable-timestamp
contract its copy never pinned. The test-local median is gone too — a
hand-rolled copy in the assertion can agree with a bug instead of
catching it.
